serial and motor control compiled successfully

+++ Apps/FlightSoftware/fprime/CubeRover/MotorControl/MotorControlComponentImpl.hpp
@@ -16,7 +16,6 @@
 #include "CubeRover/MotorControl/MotorControlComponentAc.hpp"
 #include "i2c.h"
 
-
 namespace CubeRover {
 
   #define MC_BUFFER_MAX_SIZE 16 // Maximum size of I2C buffer
@@ -185,11 +184,11 @@ namespace CubeRover {
       MCError i2cMasterReceive(i2cBASE_t *i2c, const I2cSlaveAddress sadd, const uint32_t length, uint8_t * data);
       uint32_t getSizeData(const MotorControlI2cRegId id);
 
-      size_t headerSize();
-      size_t checksumSize();
+      uint32_t headerSize();
+      uint32_t checksumSize();
       void generateChecksumTable(const ChecksumType polynomial);
-      MCError computeChecksum8(uint8_t *data, const size_t bufferLength, MotorControlChecksum *checksum);
-      MCError packTransmitBuffer(const MotorControlI2cRegId id, const uint32_t data, const size_t dataLength);
+      MCError computeChecksum8(uint8_t *data, const uint32_t bufferLength, MotorControlChecksum *checksum);
+      MCError packTransmitBuffer(const MotorControlI2cRegId id, const uint32_t data, const uint32_t dataLength);
       uint8_t checksumLookUpTable[256];
 
       Motor_tick cmToMotorTicks(const Distance_cm dist);
